TST: account for and test a small corner case
wherein files with stuff at the start of the file (not whitespace,
but other random junk) would get treated as node sequences, which caused
pyfastg to try to add a node for those sequences to the graph (which
would fail albeit with a confusing error).

Not only does this account for that case, this also should necessarily
take care of a lot of random non-FASTG files that people try and throw
at this program. Yay!