Improved validation&docs, update CI, rm skbio (#5)
* DEV: Travis --> GitHub Actions (close #3)

* REL: bump version number to 0.1.0

* STY: update based on new black vsn

* DOC: Update CI badge in README

* TST: more explicit coverage report formats

* BUG/DOC: Fix & improve edge decl err msg

Previously, I was just outputting {} instead of the actual
problematic declaration. Guess I forgot to use .format(), whoops.

* MNT: on second thought, don't use f-strings

no need to limit this to >= py 3.6 if we can avoid it

* TST: thats a lotta python

* DOC: topology note

See https://github.com/marbl/MetagenomeScope/issues/234. Thought
it'd be good to be extra explicit about this.

* DOC: clarify parsing more

* TST: no-rc test

* DOC: improve multi-colon error message

Just in case someone rolls up with a FASTG file with properties

* TST: "undefined target edge" integration test

This was already basically guaranteed, but now it's unambiguous
that the final check works.

could make the error message more descriptive (for which edges
is this attribute missing), but that'd necessitate refactoring
and not important to do that atm imo

* TST: more paranoid testing

* DEP/ENH: remove skbio dep; compute GC manually

This change makes the implicit sequence validation more explicit (we do
it), and also more lax (now uracil [U] is ok).

Closes #2.

* DOC: shorten README title

* TST: don't test on python 3.4

We could mess with the ubuntu version to try to make this work,
but i think it makes sense to prioritize support for newer python
versions. see https://github.com/actions/setup-python/issues/185

* TST: Fix 3.10 in YAML

* TST: axe py 3.5, add py 3.11

from the github actions logs:
"DEPRECATION: Python 3.5 reached the end of its life on September 13th,
2020. Please upgrade your Python as Python 3.5 is no longer maintained.
pip 21.0 will drop support for Python 3.5 in January 2021."

* TST: 3.11-dev

* DOC/ENH: tidy up declaration info extraction

- Include ^ and $ in the regex
- Document the (?P<groupname>...) stuff
- Change the name of the first capture group from node --> name
- Remove the | in the middle of the coverage set of accepted
  characters (this was a small bug -- wouldn't have done any damage
  since the attempted conversion to float would've immediately broken
  due to any |s)

* MNT: Remove +? from name and length regex

Since the sets of accepted characters for name and length don't
include "_", the minimality thing shouldn't make a difference. This
makes the code clearer. See
https://docs.python.org/3/library/re.html#index-6 for details on what
this was doing.

* TST: test bad-coverage cases

* DOC: mention invalid coverage floats in readme

* TST: test leading/trailing info not ok in decls

* TST/ENH: Add check that decl lines ends with ;

(since the spec technically allows them not to...)

* TST: another end-with-; test

* ENH/TST: Declaration consistency check!

also:

-removed the >? from the start of the main edge declaration line
 regex to make life easier
-made the (lack of) space on the line after "Parameters\n---------"
 consistent
-lotta tests

* DOC: tidy up readme some

* TST: test another decl-inconsistency case (1 line)

* DOC: better error message on inconsistent decls

wrt the node/edge debacle

* ENH: Detect and throw error if duplicate out adjs

* REL: add draft changelog

not necessarily thorough yet -- will look over again tmw

* DOC: make readme intro read better

* DOC: yassify readme intro more

* DOC: various README improvements

* TST: verify that spaces next to adj commas not ok

The error should be self-explanatory.

* ENH: hit the yoinky sploinky and disallow ~

a situation where we can be very explicit in the error message

* ENH: more blocking problematic things explicitly

* DOC: tidy up the intro of the README a bit

* DOC: Update README

- Move dependency docs up to installation
- Add dev instructions
- Link to changelog

* DOC: tidy up FASTG details in README

* DOC: impv orientation, node name formatting, etc

* DOC: Update the changelog

phew

* DOC: grammar

* DOC: grammar but again